kap8

+++ kapitel8/src/soag01/SodaCan.java
@@ -1,21 +1,27 @@
 package soag01;
+/**
+ * 
+ * @author Sofia Ågren
+ * @version2019-10-25
+ *
+ */
 
 public class SodaCan {
 
-	float height = 0.0f;
-	float radius = 0.0f;
+	double height = 0.0f;
+	double radius = 0.0f;
 	
-	public SodaCan(float height, float radius) {
+	public SodaCan(double height, double radius) {
 		this.height = height;
 		this.radius = radius;
 	}
-	public float getSurfaceArea() {
-		float area = (float)(2.0f*Math.PI*Math.pow(radius,2)+2.0f*Math.PI*radius*height);
+	public double getSurfaceArea() {
+		double area = (2.0f*Math.PI*Math.pow(radius,2)+2.0f*Math.PI*radius*height);
 		return area;
 		
 	}
-	public float getVolume() {
-		float volume = (float)(Math.PI*Math.pow(radius, 2)*height);
+	public double getVolume() {
+		double volume = (Math.PI*Math.pow(radius, 2)*height);
 		return volume;
 		
 	}
